LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/top-expected.txt


PASS Can set 'top' to CSS-wide keywords
FAIL Can set 'top' to var() references assert_equals: expected 2 but got 1
PASS Can set 'top' to the 'auto' keyword
FAIL Can set 'top' to a percent assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSUnitValue]"
FAIL Can set 'top' to a length assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSUnitValue]"
PASS Setting 'top' to a time throws TypeError
PASS Setting 'top' to an angle throws TypeError
PASS Setting 'top' to a flexible length throws TypeError
FAIL Setting 'top' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'top' to a URL throws TypeError
PASS Setting 'top' to a transform throws TypeError
